requirements.txt
pytest==8.3.3
selene>=2.0.0rc9
python-dotenv==1.0.1
allure-pytest==2.13.5
allure-python-commons==2.13.5